Add DollarlessTerminalName
This caught a bunch of bugs.
Lesson learned: Use more "tiny types". Especially when the types
are easy to confuse. I didn't think dollarless and dollared names
would be easy to confuse.
However, one key thing I overlooked was that the AST has dollared names,
and most of the code uses dollarless names.