Make `parse` return `Result<{start}, Option<Token>>`
Previously it returned`Result<{start}, Token>`.
However, if we encounter the end of input,
and the parser is not ready to accept,
then we must return an error.
But we have no token, which causes a problem.

So instead, we return an _optional_ token for the error type.
This way, we can return `None`  in the case of an
unexpected end of input.